finished working woth files

+++ src/main/java/vlad/mester/syncroflowbe/FileController.java
@@ -1,16 +1,17 @@
 package vlad.mester.syncroflowbe;
 
+import org.json.simple.JSONObject;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.nio.file.Path;
 
 public class FileController {
-    File root = new File("FileDirectory");
+    public static final File root = new File("FileDirectory");
 
     public String addDirectory(String path, String name) {
-        if (path==null)
-            path = root.getPath();
+        if (!checkPath(path))
+            return "Invalid path";
         File file = new File(path + "/" + name);
         if (!file.exists()) {
             if (file.mkdir()){
@@ -22,6 +23,8 @@ public class FileController {
     }
 
     public String removeDirectory(String path) {
+        if (!checkPath(path))
+            return "Invalid path";
         File file = new File(path);
         if(!file.exists())
             return "Directory does not exist";
@@ -36,11 +39,15 @@ public class FileController {
                 return "Directory is used";
             }
         }
+        if (file.getPath().equals(root.getPath()))
+            return "Cannot delete root directory";
         file.delete();
         return "Directory deleted";
     }
 
     public String addFile(String path, MultipartFile file) {
+        if (!checkPath(path))
+            return "Invalid path";
         Path destination = Path.of(path);
         if(!destination.toFile().exists()) {
             return "Path does not exist";
@@ -54,6 +61,8 @@ public class FileController {
     }
 
     public String removeFile(String path) {
+        if (!checkPath(path))
+            return "Invalid path";
         File file = new File(path);
         if(!file.exists())
             return "File does not exist";
@@ -67,5 +76,29 @@ public class FileController {
         }
         file.delete();
         return "File deleted";
+    }
+
+    public boolean checkPath(String path) {
+        return path.contains(root.getPath());
+    }
+
+    public JSONObject getDirectoryStructure(String path) {
+        if (!checkPath(path))
+            return null;
+        File file = new File(path);
+        if(!file.exists())
+            return null;
+        JSONObject directory = new JSONObject();
+        directory.put("name", file.getName());
+        directory.put("path", file.getPath());
+        directory.put("isDirectory", file.isDirectory());
+        if (file.isDirectory()) {
+            JSONObject[] children = new JSONObject[file.list().length];
+            for (int i = 0; i < file.list().length; i++) {
+                children[i] = getDirectoryStructure(file.getPath() + "/" + file.list()[i]);
+            }
+            directory.put("children", children);
+        }
+        return directory;
     }
 }
